+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/process/DefaultKlumPhase.java
@@ -33,6 +33,7 @@ public enum DefaultKlumPhase implements KlumPhase {
     /** The creation phase is not encountered in the PhaseDriver, it handles the actual creation of the objects. */
     CREATE(0),
     APPLY_LATER(1),
+    EARLY_VALIDATE(5),
     /** Phase for automatic creation of missing objects, usually from annotations. */
     AUTO_CREATE(10),
     OWNER(15),