+++ CHANGES.md
@@ -4,10 +4,12 @@
 - Validation improvements (see [Validation](https://github.com/klum-dsl/klum-ast/wiki/Validation) [#395](https://github.com/klum-dsl/klum-ast/issues/395))
   - Validation-phase is split into validation and verify phases.
   - Provided new (preliminary) methods to explicitly create validation issues. This allows validation issues to be created in earlier phases, as well as multiple issues in a single validation/lifecycle method.
-  - further validation issues can explicitly be suppressed for specific fields
+  - further validation issues can explicitly be suppressed for specific fields and specific maximum levels
   - New annotation `@Optional` as alias for `@Validate(Validate.Ignore)`
   - VerifyPhase can now be skipped using system property `klum.validation.skipVerify`
   - Results of a complete structure can be retrieved using `Validator.getValidationResultsFromStructure(Object)` or verified later using `Validator.verifyStructure(Object)`
+  - Deprecation checks now run in the new early validation phase and check only for manually set values, not values created by later phases
+  - new annotation `@Notify` to raise issues if a field is set or unset in the early validation phase.
 
 ## Bugfixes
 - `StructureUtil.getPathOfFieldContaining()` and therefore `@Role` fields ignored fields where the value was actually a subclass of the field type.